Chat: Combat Repeat db conversion
On 6.xx version, it is assumed that the value of 0 disabled the feature.
On v7, there is now a toggle for Combat Repeat, so apply that conversion here.
There is also min value of 2, but value == 1 wont be migrated since there is no breaking of behaviour from previous profile.

The conversion function only runs once  by E:DBConversions() when `E.db.dbConverted ~= E.version`, so this will need manual patching for users that have run Dev prior to this commit